# testFuncional/swagger-tarjeta-credito.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8081","basePath":"/","tags":[{"name":"#TarjetaCredito","description":"Tarjeta Credito Controller"}],"paths":{"/tarjetacredito/actualizar":{"put":{"tags":["#TarjetaCredito"],"summary":"Actualizar Tarjeta de Credito","operationId":"actualizarTarjetaCreditoUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"tarjetaCredito","description":"tarjetaCredito","required":true,"schema":{"$ref":"#/definitions/TarjetaCredito"}}],"responses":{"200":{"description":"Tarjeta de Credito actualizada."},"201":{"description":"Created"},"400":{"description":"Datos de entrada incorrectos."},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Error inesperado."}},"deprecated":false}},"/tarjetacredito/consultar":{"get":{"tags":["#TarjetaCredito"],"summary":"Consultar Tarjetas de Credito","operationId":"consultarTarjetaCreditosUsingGET","produces":["*/*"],"responses":{"200":{"description":"Tarjetas de Credito consultadas.","schema":{"type":"array","items":{"$ref":"#/definitions/TarjetaCredito"}}},"400":{"description":"Datos de entrada incorrectos."},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Error inesperado."}},"deprecated":false}},"/tarjetacredito/consultar/{id}":{"get":{"tags":["#TarjetaCredito"],"summary":"Consultar Tarjeta de Credito por ID","operationId":"consultarTarjetaCreditoPorIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Tarjeta de Credito consultada.","schema":{"$ref":"#/definitions/TarjetaCredito"}},"400":{"description":"Datos de entrada incorrectos."},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Error inesperado."}},"deprecated":false}},"/tarjetacredito/eliminar/{id}":{"delete":{"tags":["#TarjetaCredito"],"summary":"Eliminar Tarjeta de Credito","operationId":"eliminarTarjetaCreditoUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Tarjeta de Credito eliminada."},"204":{"description":"No Content"},"400":{"description":"Datos de entrada incorrectos."},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"500":{"description":"Error inesperado."}},"deprecated":false}},"/tarjetacredito/guardar":{"post":{"tags":["#TarjetaCredito"],"summary":"Crear Tarjeta de Credito","operationId":"guardarTarjetaCreditoUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"tarjetaCredito","description":"tarjetaCredito","required":true,"schema":{"$ref":"#/definitions/TarjetaCredito"}}],"responses":{"200":{"description":"Tarjeta de Credito creada."},"201":{"description":"Created"},"400":{"description":"Datos de entrada incorrectos."},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Error inesperado."}},"deprecated":false}},"/tarjetacredito/persona/guardar":{"post":{"tags":["#TarjetaCredito"],"summary":"Guardar Tarjetas de Credito asociadas a una persona","operationId":"guardarTarjetasCreditoUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"tarjetasCredito","description":"tarjetasCredito","required":true,"schema":{"type":"array","items":{"$ref":"#/definitions/TarjetaCredito"}}}],"responses":{"200":{"description":"Registros creados."},"201":{"description":"Created"},"400":{"description":"Datos de entrada incorrectos."},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Error inesperado."}},"deprecated":false}}},"definitions":{"TarjetaCredito":{"type":"object","required":["banco","fechaVencimiento","idPersona","numero"],"properties":{"banco":{"type":"string","description":"Fecha de Vencimiento de la tarjeta"},"fechaVencimiento":{"type":"string","format":"date-time","description":"Fecha de Vencimiento de la tarjeta"},"id":{"type":"integer","format":"int64"},"idPersona":{"type":"integer","format":"int64","description":"ID de la persona"},"numero":{"type":"string","description":"Numero de la tarjeta"}},"title":"TarjetaCredito"}}}